+++ src/pwg-global.h
@@ -208,6 +208,20 @@ PWG_API
 gboolean pwg_global_is_client(PwgGlobal *self);
 
 /**
+ * pwg_global_is_device:
+ * @self: a PipeWire global descriptor.
+ *
+ * Checks whether this global describes a PipeWire device.
+ *
+ * Returns: whether this global is a PipeWire device.
+ *
+ * Since: 0.1
+ * Stability: Unstable
+ */
+PWG_API
+gboolean pwg_global_is_device(PwgGlobal *self);
+
+/**
  * pwg_global_is_port:
  * @self: a PipeWire global descriptor.
  *